moved the decoding user from JWT function to context
